incident-agentic-workflow/requirements.txt
# -------------------------
# Core (always required)
# -------------------------
python-dotenv>=1.0.0

# -------------------------
# LangGraph (deterministic control plane)
# -------------------------
langgraph>=0.0.40
# Align with langchain_openai expectations
langchain-core>=0.2.10

# Optional: LLM provider if you later plug in real models
# openai>=1.0.0

# -------------------------
# AutoGen (non-deterministic / exploratory)
# -------------------------
pyautogen>=0.2.0
autogen>=0.2.0

# -------------------------
# Testing
# -------------------------
pytest>=7.4.0

langchain-openai>=0.1.6
langchain-community>=0.2.10
openai>=1.66.2
faiss-cpu
pydantic
autogen